TB_Detection/requirements.txt
tensorflow>=2.16.0
numpy>=1.24.0
pandas>=1.5.2
matplotlib>=3.6.2
seaborn>=0.12.1
scikit-learn>=1.2.0
streamlit>=1.16.0
opencv-python-headless>=4.6.0.66
pillow>=9.3.0
albumentations>=1.3.0
kaggle>=1.5.12
boto3>=1.26.13
pytest>=7.2.0